Move inspect API files to their own dir.
