+++ packages/breadboard/tests/inspector/describe.ts
@@ -6,7 +6,7 @@
 
 import test from "ava";
 
-import { inspectableGraph } from "../../src/inspector/mutable-graph.js";
+import { inspectableGraph } from "../../src/inspector/graph/mutable-graph.js";
 import { createLoader } from "../../src/loader/index.js";
 import { GraphDescriptor, Schema } from "../../src/types.js";
